Created Location class and removed locaiton from Resource
