+++ LiveSched/src/main/java/dev/coms4156/project/livesched/Resource.java
@@ -4,49 +4,25 @@ import java.time.LocalDateTime;
 
 /**
  * Represents a resource that can be assigned to a task.
- * This class stores the id, availability, and location of the resource.
+ * This class stores the id and availability of the resource.
  */
 public class Resource {
   private final String resourceId;
   private LocalDateTime availableFrom;
-  private double latitude;
-  private double longitude;
 
   /**
    * Constructs a new Resource object with the given parameters.
    *
    * @param resourceId  the unique ID of the resource
-   * @param latitude    the latitude of the resource's location
-   * @param longitude   the longitude of the resource's location
    * @throws IllegalArgumentException if the {@code resourceId} is null or empty,
    *                                  or if the latitude or longitude is out of bounds
    */
-  public Resource(String resourceId, double latitude, double longitude) {
+  public Resource(String resourceId) {
     if (resourceId == null || resourceId.trim().isEmpty()) {
       throw new IllegalArgumentException("Resource ID cannot be null or empty.");
     }
-    validateLatLong(latitude, longitude);
-
     this.resourceId = resourceId;
     this.availableFrom = LocalDateTime.now(); // Initially available now
-    this.latitude = latitude;
-    this.longitude = longitude;
-  }
-
-  /**
-   * Validates that the given latitude and longitude are within their valid ranges.
-   *
-   * @param latitude  the latitude
-   * @param longitude the longitude
-   * @throws IllegalArgumentException if the latitude or longitude is out of bounds
-   */
-  private void validateLatLong(double latitude, double longitude) {
-    if (latitude < -90 || latitude > 90) {
-      throw new IllegalArgumentException("Latitude must be between -90 and 90.");
-    }
-    if (longitude < -180 || longitude > 180) {
-      throw new IllegalArgumentException("Longitude must be between -180 and 180.");
-    }
   }
 
   /**
@@ -91,19 +67,6 @@ public class Resource {
   }
 
   /**
-   * Updates the location of the resource.
-   *
-   * @param latitude  the new latitude of the resource's location
-   * @param longitude the new longitude of the resource's location
-   * @throws IllegalArgumentException if the latitude or longitude is out of bounds
-   */
-  public void updateLocation(double latitude, double longitude) {
-    validateLatLong(latitude, longitude);
-    this.latitude = latitude;
-    this.longitude = longitude;
-  }
-
-  /**
    * Gets the ID of the Resource.
    *
    * @return A {@code String} of resourceId
@@ -128,10 +91,6 @@ public class Resource {
       throw new IllegalArgumentException("Time available from cannot be null.");
     }
     this.availableFrom = availableFrom;
-  }
-
-  public String getLocation() {
-    return latitude + ", " + longitude;
   }
 
 }
